pyproject.toml
[project]
name = "ppt3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.7",
    "apscheduler>=3.11.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "paddleocr==2.9.0",
    "paddlepaddle==3.0.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.3",
    "setuptools>=80.8.0",
    "wheel>=0.45.1",
]
